+++ tests/integrations/torch/train_test.py
@@ -20,7 +20,11 @@ class TestTrainStep(TangoTestCase):
                 "test_fixtures.integrations.common",
                 "test_fixtures.integrations.torch",
             ],
-            overrides="" if with_validation else json.dumps({"steps.train.validation_split": None}),
+            overrides=""
+            if with_validation
+            else json.dumps(
+                {"steps.train.validation_split": None, "steps.train.validate_every": None}
+            ),
         )
         assert (result_dir / "train" / "data.pt").is_file()
         assert (result_dir / "train" / "work" / "state_worker0.pt").is_file()
@@ -33,7 +37,13 @@ class TestTrainStep(TangoTestCase):
                 "test_fixtures.integrations.common",
                 "test_fixtures.integrations.torch",
             ],
-            overrides=json.dumps({"steps.train.train_steps": None, "steps.train.train_epochs": 2}),
+            overrides=json.dumps(
+                {
+                    "steps.train.train_steps": None,
+                    "steps.train.train_epochs": 2,
+                    "steps.train.validate_every": None,
+                }
+            ),
         )
         assert (result_dir / "train" / "data.pt").is_file()