+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.5.0](https://github.com/allenai/tango/releases/tag/v0.5.0) - 2022-02-09
+
 ### Added
 
 - Added `TrainingEngine` abstraction to torch integration.